pyproject.toml
[tool.poetry]
name = "hw-drf"
version = "0.1.0"
description = ""
authors = ["KonstantinMikheev <mikhey17@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.2"
djangorestframework = "^3.15.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pillow = "^11.0.0"
black = "^24.10.0"
django-stubs = "^5.1.0"
django-filter = "^24.3"
djangorestframework-simplejwt = "^5.3.1"
ipython = "^8.29.0"
isort = "^5.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
